Broke up main view into a number of sub views. Added model.
